pbr/ibl working again

+++ src/vulkan/Window.h
@@ -9,7 +9,7 @@ struct Window {
 
 	void updateWindowSize() const;
 
-	bool throttleIfWindowUnfocused(int sleepMs) const;
+	bool throttleIfWindowUnfocused(double sleepMs) const;
 
 	void initWindow(const uint32_t width, const uint32_t height);
 	void cleanupWindow() const;
